Fix return types

+++ src/storage/storage.h
@@ -13,7 +13,7 @@ storage_channel_t* storage_open(
 storage_channel_t* storage_open_fd(
         storage_io_common_fd_t fd);
 
-size_t storage_readv_internal(
+int32_t storage_readv_internal(
         storage_channel_t *channel,
         storage_io_common_iovec_t *iov,
         size_t iov_nr,
@@ -32,7 +32,7 @@ bool storage_read(
         size_t buffer_len,
         off_t offset);
 
-size_t storage_read_try(
+int32_t storage_read_try(
         storage_channel_t *channel,
         char *buffer,
         size_t buffer_len,
